refactor(commands): render_composition_manifest
Extract to a function so that it can be re-used for trust of included
remote environments by `flox activate`.

This is already "integration" tested by `flox list` so I'm not sure how
I feel about writing additional unit tests right now.
